rotas das rooms

+++ database/migrations/2019_07_31_174349_create_posts_in_rooms.php
@@ -15,7 +15,7 @@ class CreatePostsInRooms extends Migration
     {
         Schema::create('posts_in_rooms', function (Blueprint $table) {
             $table->increments('id');
-            $table->unsignedInteger('post_id');
+            $table->unsignedBigInteger('post_id');
             $table->unsignedInteger('room_id');
             $table->foreign('post_id')->references('id')->on('posts')->notnull();
             $table->foreign('room_id')->references('id')->on('rooms')->notnull();
